⚡️(performance) fix rendering chat messages
Extracted message rendering into a memoized ChatMessage component to prevent unnecessary re-renders when typing in the input.
